# Stop inheriting `unittest.TestCase` for performance