… tests/cupy_tests/math_tests/test_sumprod.py
… tests/cupy_tests/random_tests/test_distributions.py
… tests/cupy_tests/test_cublas.py
… tests/cupy_tests/test_cusolver.py
… tests/cupy_tests/test_cusparse.py
… tests/cupyx_tests/scipy_tests/fft_tests/test_fft.py
… tests/cupyx_tests/scipy_tests/signal_tests/test_signaltools.py
… tests/cupyx_tests/scipy_tests/sparse_tests/test_construct.py
… tests/cupyx_tests/scipy_tests/sparse_tests/test_coo.py
… tests/cupyx_tests/scipy_tests/sparse_tests/test_csc.py
… tests/cupyx_tests/scipy_tests/sparse_tests/test_csr.py
… tests/cupyx_tests/scipy_tests/sparse_tests/test_index.py