`cd tests; for i in *test*/*test*; do echo $i; pytest -q --collect-only $i | tail -n 1; done`

||before|after|
|-|-|-|
|cupy_tests/binary_tests|9 tests collected in 0.70s|9 tests collected in 0.75s|
|cupy_tests/core_tests|2357 tests collected in 13.82s|2357 tests collected in 8.13s|
|cupy_tests/creation_tests|415 tests collected in 2.07s|415 tests collected in 1.86s|
|cupy_tests/cuda_tests|834 tests collected in 5.95s|834 tests collected in 1.90s|
|cupy_tests/fft_tests|2711 tests collected in 1.90s|2711 tests collected in 1.87s|
|cupy_tests/functional_tests|67 tests collected in 0.75s|67 tests collected in 0.72s|
|cupy_tests/indexing_tests|384 tests collected in 2.44s|384 tests collected in 2.30s|
|cupy_tests/io_tests|37 tests collected in 0.67s|37 tests collected in 0.68s|
|cupy_tests/lib_tests|484 tests collected in 7.79s|484 tests collected in 1.41s|
|cupy_tests/linalg_tests|1327 tests collected in 8.20s|1327 tests collected in 3.74s|
|cupy_tests/logic_tests|788 tests collected in 2.40s|788 tests collected in 0.89s|
|cupy_tests/manipulation_tests|479 tests collected in 1.39s|479 tests collected in 1.43s|
|cupy_tests/math_tests|9445 tests collected in 11.24s|9445 tests collected in 5.37s|
|cupy_tests/misc_tests|22 tests collected in 0.66s|22 tests collected in 0.66s|
|cupy_tests/padding_tests|75 tests collected in 1.45s|75 tests collected in 1.29s|
|cupy_tests/polynomial_tests|55 tests collected in 0.74s|55 tests collected in 0.74s|
|cupy_tests/prof_tests|8 tests collected in 0.72s|8 tests collected in 0.69s|
|cupy_tests/random_tests|1880 tests collected in 22.85s|1880 tests collected in 6.68s|
|cupy_tests/sorting_tests|412 tests collected in 3.45s|412 tests collected in 3.48s|
|cupy_tests/statistics_tests|662 tests collected in 7.40s|662 tests collected in 1.37s|
|cupy_tests/test_cublas.py|2436 tests collected in 12.74s|2436 tests collected in 0.94s|
|cupy_tests/test_cudnn.py|2420 tests collected in 25.46s|2420 tests collected in 0.95s|
|cupy_tests/test_cusolver.py|648 tests collected in 4.29s|648 tests collected in 0.87s|
|cupy_tests/test_cusparse.py|1027 tests collected in 20.64s|1027 tests collected in 1.30s|
|cupy_tests/test_cutensor.py|75 tests collected in 1.05s|75 tests collected in 0.67s|
|cupy_tests/test_init.py|11 tests collected in 0.65s|11 tests collected in 0.65s|
|cupy_tests/test_ndim.py|12 tests collected in 0.64s|12 tests collected in 0.66s|
|cupy_tests/test_numpy_interop.py|9 tests collected in 0.70s|9 tests collected in 0.65s|
|cupy_tests/test_type_routines.py|34 tests collected in 0.76s|34 tests collected in 0.71s|
|cupy_tests/testing_tests|289 tests collected in 1.86s|289 tests collected in 1.80s|
|cupyx_tests/fallback_mode_tests|107 tests collected in 1.48s|107 tests collected in 1.47s|
|cupyx_tests/jit_tests|43 tests collected in 0.65s|43 tests collected in 0.72s|
|cupyx_tests/linalg_tests|32 tests collected in 0.71s|32 tests collected in 0.71s|
|cupyx_tests/scipy_tests|69228 tests collected in 238.70s (0:03:58)|68736 tests collected in 18.04s|
|cupyx_tests/test_cupyx.py|2 tests collected in 0.79s|2 tests collected in 0.65s|
|cupyx_tests/test_lapack.py|70 tests collected in 0.96s|70 tests collected in 0.88s|
|cupyx_tests/test_optimize.py|9 tests collected in 1.20s|9 tests collected in 1.21s|
|cupyx_tests/test_pinned_array.py|76 tests collected in 0.80s|76 tests collected in 0.79s|
|cupyx_tests/test_rsqrt.py|1 test collected in 0.64s|1 test collected in 0.65s|
|cupyx_tests/test_runtime.py|2 tests collected in 0.65s|2 tests collected in 0.65s|
|cupyx_tests/test_scatter.py|3 tests collected in 0.65s|3 tests collected in 0.66s|
|cupyx_tests/test_time.py|9 tests collected in 0.67s|9 tests collected in 0.64s|
|cupyx_tests/tools_tests|9 tests collected in 0.68s|9 tests collected in 0.66s|
|example_tests/example_test.py|no tests collected in 0.03s|no tests collected in 0.03s|
|example_tests/test_custom_struct.py|3 tests collected in 0.04s|3 tests collected in 0.03s|
|example_tests/test_finance.py|3 tests collected in 0.66s|3 tests collected in 0.67s|
|example_tests/test_gemm.py|1 test collected in 0.04s|1 test collected in 0.03s|
|example_tests/test_gmm.py|2 tests collected in 0.72s|2 tests collected in 0.77s|
|example_tests/test_kmeans.py|3 tests collected in 0.76s|3 tests collected in 0.71s|
|install_tests/test_build.py|3 tests collected in 0.04s|3 tests collected in 0.04s|
|install_tests/test_utils.py|2 tests collected in 0.04s|2 tests collected in 0.03s|