@leofang 2021-08-04
@toslunar This is very cool! Do you have insight as for why not subclassing `unittest.TestCase` can make such a huge performance difference? Also, not all of the cases have the collection time reduced, though it's not getting a lot worse either so we should go for it.