@toslunar 2021-08-05
Sorry, I forgot to write the background.

`cupy.testing.parametrize` is slow if we cannot use `pytest.mark.parametrize`. PyTest does not guarantee that `pytest.mark.parametrize` works with `unittest.TestCase` (https://docs.pytest.org/en/stable/unittest.html#pytest-features-in-unittest-testcase-subclasses), while simple cases seem working totally fine.  The decision in #4192 was
> For now, the trigger is to stop inheriting unittest.TestCase.